#54 オッズ一覧、三連複と三連単を作成

+++ KmyKeiba.Data/Entities/ExactaOdds.cs
@@ -22,7 +22,7 @@ namespace KmyKeiba.JVLink.Entities
 
       public short HorseNumber2 { get; init; }
 
-      public short Odds { get; init; }
+      public uint Odds { get; init; }
     }
 
     public static ExactaOdds FromJV(JVData_Struct.JV_O4_ODDS_UMATAN odds)
@@ -45,7 +45,7 @@ namespace KmyKeiba.JVLink.Entities
           continue;
         }
 
-        short.TryParse(data.Odds, out short oval);
+        uint.TryParse(data.Odds, out uint oval);
 
         od.Odds.Add(new OddsData
         {
