+++ KmyKeiba.Data/Entities/TrifectaOdds.cs
@@ -24,7 +24,7 @@ namespace KmyKeiba.JVLink.Entities
 
       public short HorseNumber3 { get; init; }
 
-      public short Odds { get; init; }
+      public uint Odds { get; init; }
     }
 
     public static TrifectaOdds FromJV(JVData_Struct.JV_O6_ODDS_SANRENTAN odds)
@@ -48,7 +48,7 @@ namespace KmyKeiba.JVLink.Entities
           continue;
         }
 
-        short.TryParse(data.Odds, out short oval);
+        uint.TryParse(data.Odds, out uint oval);
 
         od.Odds.Add(new OddsData
         {
